requirements.txt
asgiref==3.6.0
Django==3.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-mptt==0.14.0
Pillow==9.3.0
pytz==2022.7
sqlparse==0.4.3
